Created pages folder + profile, moved menu to app
